Gen2 Charger Control code
Started adding functionality to control the Charger over CAN.  In the TesLorean, the charger will be configured to operate most often in AutoCharging mode (as coded by Tom Debree), but the Trip Computer will also be able to start scheduled charging (such as in the middle of the night).  For a scheduled charge, the Trip computer instructs the charger to drop out of auto mode, and then instructs the charger to start charging.